ユーザーごとで表示絞りこみ

+++ edit.php
@@ -100,8 +100,9 @@ if (isset($_GET['id'])) {
       </form>
     </div>
     <a href="date.php">カレンダーへ</a>
-    <a href="delete.php?id=<?php echo $shcedule->getId() ?>" class="">削除</a>
+    <a onclick="return confirm('本当に削除してよろしいですか?')" href="delete.php?id=<?php echo $shcedule->getId() ?>" class="">削除</a>
   </div>
+  <script src="check.js"></script>
 
 </body>
